+++ src/Razor/test/Microsoft.VisualStudio.Razor.Integration.Test/InProcess/EditorInProcess_LightBulb.cs
@@ -46,7 +46,7 @@ namespace Microsoft.VisualStudio.Razor.Integration.Test.InProcess
             }
 
             var session = broker.GetSession(view);
-            if (session == null || !session.IsExpanded)
+            if (session is null || !session.IsExpanded)
             {
                 return false;
             }
