+++ src/Razor/test/Microsoft.VisualStudio.Razor.Integration.Test/LanguageNames.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT license. See License.txt in the project root for license information.
+
+namespace Microsoft.VisualStudio.Razor.Integration.Test
+{
+    public static class LanguageNames
+    {
+        public const string Razor = "Razor";
+        public const string CSharp = "CSharp";
+    }
+}
